Enable use-cases to be linked-to
Enable use-cases to be linked to by using markdown headers to provide
HTML anchor points. Parts of the text have been updated to remove the
use of markdown as styling which has resulted in most headers being
brought up one level. Titles have been changed to remove punctuation
which affects how headers are rendered. This change means that a
table of contents can be more easily rendered using the semantics
encoded in the markdown.